+++ tests/test_converter.py
@@ -9,6 +9,7 @@ from typing import (
     Sequence,
     Set,
     Tuple,
+    Type,
     Union,
 )
 
@@ -18,7 +19,11 @@ from hypothesis import HealthCheck, assume, given, settings
 from hypothesis.strategies import booleans, just, lists, one_of, sampled_from
 
 from cattrs import BaseConverter, Converter, UnstructureStrategy
-from cattrs.errors import ClassValidationError, ForbiddenExtraKeysError
+from cattrs.errors import (
+    ClassValidationError,
+    ForbiddenExtraKeysError,
+    StructureHandlerNotFoundError,
+)
 from cattrs.gen import make_dict_structure_fn, override
 
 from ._compat import is_py39_plus, is_py310_plus
@@ -707,3 +712,58 @@ def test_annotated_with_typing_extensions_attrs():
 
     structured = converter.structure(raw, Outer)
     assert structured == Outer(Inner(2), [Inner(2)], Inner(2))
+
+
+def test_unstructure_fallbacks(converter_cls: Type[BaseConverter]):
+    """Unstructure fallback factories work."""
+
+    class Test:
+        """Unsupported by default."""
+
+    c = converter_cls()
+
+    assert isinstance(c.unstructure(Test()), Test)
+
+    c = converter_cls(
+        unstructure_fallback_factory=lambda _: lambda v: v.__class__.__name__
+    )
+    assert c.unstructure(Test()) == "Test"
+
+
+def test_structure_fallbacks(converter_cls: Type[BaseConverter]):
+    """Structure fallback factories work."""
+
+    class Test:
+        """Unsupported by default."""
+
+    c = converter_cls()
+
+    with pytest.raises(StructureHandlerNotFoundError):
+        c.structure({}, Test)
+
+    c = converter_cls(structure_fallback_factory=lambda _: lambda v, _: Test())
+    assert isinstance(c.structure({}, Test), Test)
+
+
+def test_fallback_chaining(converter_cls: Type[BaseConverter]):
+    """Converters can be chained using fallback hooks."""
+
+    class Test:
+        """Unsupported by default."""
+
+    parent = converter_cls()
+
+    parent.register_unstructure_hook(Test, lambda _: "Test")
+    parent.register_structure_hook(Test, lambda v, _: Test())
+
+    # No chaining first.
+    child = converter_cls()
+    with pytest.raises(StructureHandlerNotFoundError):
+        child.structure(child.unstructure(Test()), Test)
+
+    child = converter_cls(
+        unstructure_fallback_factory=parent._unstructure_func.dispatch,
+        structure_fallback_factory=parent._structure_func.dispatch,
+    )
+
+    assert isinstance(child.structure(child.unstructure(Test()), Test), Test)